change the way the critical message is copied
When one of the messages provided by the attacker is
selected, depending on the value of the secret bit, we make
sure that those messages are of the same length and then we
make a copy of one of them (not, as previously, take the
reference to one of them). This should allow Joana to see
that the length is not high, but only the content.